:hammer: chore: update values utils
